Use minimal signing certificate to shave off 142 bytes
This reduces APK size by 142 bytes, to 678 bytes, by shrinking the
APK's signature (APK Signature Scheme v2) from 603 to 461 bytes.

The APK's Signature Scheme v2 signature includes the signing
certificate in uncompressed form. Most information contained in the
certificate is not needed by Android or the APK signing tools. This
change thus switches to a hand-crafted minimalistic certificate which
is good enough for Android and apksigner.